EXPERIMENTAL: Add raw packet interface and mctp-tun util
This change adds an experimental utility to forward raw packets between
a tun device and an astlpc mctp binding. This allows us to use a
userspace astlpc binding with the in-kernel MCTP stack.

To do this, we need to extend the core a little to povide a "raw" packet
interface, where packets are sent/recieved directly to the binding,
bypassing any packetisaton/reassembly/etc.

Completely untested; I may be using the tun ABI incorrectly, and I have
no astlpc hardware to try.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>
Change-Id: I2bd48fc21e1a2138f853220e8698ae14550c8193